+++ fuzz_test.go
@@ -69,9 +69,9 @@ func FuzzDraftInvalidVectorJSONLoader(f *testing.F) {
 
 func FuzzProtocolConsistency(f *testing.F) {
 	f.Add([]byte("sid"), []byte("ctx"), []byte("ADa"), []byte("ADb"))
-	f.Add([]byte{}, []byte{}, []byte{}, []byte{})
+	f.Add([]byte("sid2"), []byte{}, []byte{}, []byte{})
 	f.Fuzz(func(t *testing.T, sid, ctx, ada, adb []byte) {
-		if len(sid) > 1024 || len(ctx) > 1024 || len(ada) > 1024 || len(adb) > 1024 {
+		if len(sid) == 0 || len(sid) > 1024 || len(ctx) > 1024 || len(ada) > 1024 || len(adb) > 1024 {
 			t.Skip()
 		}
 		initCfg := Config{
@@ -108,9 +108,9 @@ func FuzzProtocolConsistency(f *testing.F) {
 
 func FuzzProtocolMismatch(f *testing.F) {
 	f.Add([]byte("sid"), []byte("ctx"), []byte("ADa"), []byte("ADb"))
-	f.Add([]byte{}, []byte{}, []byte{}, []byte{})
+	f.Add([]byte("sid2"), []byte{}, []byte{}, []byte{})
 	f.Fuzz(func(t *testing.T, sid, ctx, ada, adb []byte) {
-		if len(sid) > 1024 || len(ctx) > 1024 || len(ada) > 1024 || len(adb) > 1024 {
+		if len(sid) == 0 || len(sid) > 1024 || len(ctx) > 1024 || len(ada) > 1024 || len(adb) > 1024 {
 			t.Skip()
 		}
 		initCfg := Config{